% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{kNN}
\alias{kNN}
\title{Nearest neighbor}
\usage{
kNN(model, k)
}
\arguments{
\item{model}{nearest neighbor ranking model}

\item{k}{number of the nearest neighbors to consider}
}
\value{
a vector of length of \code{model}
}
\description{
An auxiliary  function to find the nearest neighbors from the distance matrix
}
\details{
This function is applied to find the nearest neighbors in the distance matrix.
}

