% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr.nom}
\alias{lr.nom}
\title{Synthetic data for label ranking experiments.}
\format{A matrix of 20 discrete values and 2 columns}
\usage{
lr.nom
}
\description{
Synthetic data for label ranking experiments.
}
\keyword{datasets}

