% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr.num}
\alias{lr.num}
\title{Synthetic data for label ranking experiments.}
\format{A matrix of 20 numeric values and 2 columns}
\usage{
lr.num
}
\description{
Synthetic data for label ranking experiments.
}
\keyword{datasets}

