% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nbr.R
\name{model_nbr}
\alias{model_nbr}
\title{A naive Bayes label ranking model}
\usage{
model_nbr(x, y, n = 1)
}
\arguments{
\item{x}{is  \code{n x p} matrix of \code{n} observations and \code{p} training attributes and can have continuous or nominal values.}

\item{y}{is \code{n x j} matrix of training rankings (permutations).}

\item{n}{is a parameter of 'memory'; that is, how fast past gets forgotten. (see details of \link{time_weights}).}
}
\value{
a list of size two: prior and conditional label ranking probabilities.
}
\description{
This is an auxiliary function to build a necessary inputs to predict rankings.
}

