% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_nbr.R
\name{time_weights}
\alias{time_weights}
\title{Weights for timing}
\usage{
time_weights(x, n)
}
\arguments{
\item{x}{a scalar of timing periods.}

\item{n}{is a parameter of 'memory' of how fast the past gets forgotten.}
}
\value{
a vector of values.
}
\description{
This function calculates the diminishing weights for label ranking probabilities in case of timing nature of rankings.
}
\details{
Sometimes rankings have a 'timing' component (for example, weekly sport teams standing) and a recent event can be more important than the past. The model can take advantage of this difference in importance by weighting the ranking probabilities. The weights are calculated using an exponential function . In case of \code{n=1}, weights are a unitary vector of length \code{n}; thus, no time nature in rankings.
}

